Split FormX interface from MobX implementation
