+++ lib/flutter_formx.dart
@@ -1,5 +1,5 @@
 export 'src/form/form_x_field.dart';
-export 'src/form/formx.dart';
+export 'src/form/formx_mobx.dart';
 export 'src/validator/required_field_validator.dart';
 export 'src/validator/validator.dart';
 export 'src/validator/validator_result.dart';
