+++ lib/src/form/formx.dart
@@ -1,5 +1,4 @@
 import 'package:flutter_formx/src/form/form_x_field.dart';
-import 'package:mobx/mobx.dart';
 
 /// [FormX] is a helper class to handle forms. [T] stands for the type used to identify the fields
 /// such as an enum or a string to later access each of its fields.
@@ -15,61 +14,42 @@ import 'package:mobx/mobx.dart';
 /// The third and final step is to call [validateForm] to validate all fields and use the computed
 /// value [isFormValid] to show a submit button as enabled or disabled and verify the status of the
 /// form.
-mixin FormX<T> {
+abstract class FormX<T> {
   /// The map of fields, along with all of their properties
-  @observable
-  final ObservableMap<T, FormXField> inputMap = <T, FormXField>{}.asObservable();
+  final Map<T, FormXField> inputMap = {};
 
-  /// The computed value of the form, true if all fields are valid, false otherwise
-  @computed
-  bool get isFormValid => inputMap.values.every((element) => element.isValid);
+  /// The computed value of the form's validation, true if all fields are valid, false otherwise
+  bool get isFormValid;
 
   /// Sets up the form with the given inputs
   ///
   /// This method should be called when starting the viewModel and it already validates the form
   /// without applying any error messages.
-  Future<void> setupForm(Map<T, FormXField> inputs) {
-    inputMap.addAll(inputs);
-    return validateForm(softValidation: true);
-  }
+  Future<void> setupForm(Map<T, FormXField> inputs);
 
   /// Updates the value of the field and validates it, updating the computed variable [isFormValid].
   /// When [softValidation] is true, it doesn't add errors messages to the fields, but updates the
   /// computed variable [isFormValid] which can be used to show a submit button as enabled or disabled
-  @action
   Future<void> updateAndValidateField(
     dynamic newValue,
     T type, {
     bool softValidation = false,
-  }) async {
-    inputMap[type] = await inputMap[type]!
-        .updateValue(newValue)
-        .validateItem(softValidation: softValidation);
-  }
+  });
 
-  /// Updates the value of the field without validating it, this does not update the computed variable [isFormValid]
-  @action
-  void updateField(dynamic newValue, T type) {
-    inputMap[type] = inputMap[type]!.updateValue(newValue);
-  }
+  /// Updates the value of the field without validating it, this does not update the
+  /// computed variable [isFormValid]
+  void updateField(dynamic newValue, T type);
 
   /// Validates all fields in the form
   ///
   /// Returns bool indicating if the form is valid and when [softValidation] is true,
   /// it doesn't add errors messages to the fields, but updates the computed variable [isFormValid]
   /// which can be used to show a submit button as enabled or disabled
-  @action
-  Future<bool> validateForm({bool softValidation = false}) async {
-    await Future.forEach(inputMap.keys, (type) async {
-      inputMap[type] =
-          await inputMap[type]!.validateItem(softValidation: softValidation);
-    });
-    return isFormValid;
-  }
+  Future<bool> validateForm({bool softValidation = false});
 
   /// Returns the value of the field
-  V getFieldValue<V>(dynamic key) => inputMap[key]?.value as V;
+  V getFieldValue<V>(dynamic key);
 
   /// Returns the error message of the field
-  String? getFieldErrorMessage(dynamic key) => inputMap[key]?.errorMessage;
+  String? getFieldErrorMessage(dynamic key);
 }
